use a session manager without calling employ()
Warning: array_shift() expects parameter 1 to be array, null given in EpiSession.php on line 58

If you initialize one, and only one, of the session implementations, you should not have to call employ().

Epi::init('session-apc');
getSession()->set('this','that');

that code would generate a warning because employ is called without an argument so it returns null. array_shift() is then called on $employ which causes the warning. I think this could be cleaned up more, but this is a quick fix.